% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_names}
\alias{check_names}
\title{Check names for the code < block < chapter}
\usage{
check_names(all_info)
}
\arguments{
\item{all_info}{A vector with code < block < chapter
or block < chapter or just chapter}
}
\value{
A vector with homogenous names
}
\description{
Renames all of the entries in \code{all_info} vector
as code, block or chapter
}
\keyword{internal}
