% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{code_place}
\alias{code_place}
\title{Verify and Convert Place of Occurrence External Cause Code}
\usage{
code_place(input)
}
\arguments{
\item{input}{Either a number or a string representing the place of occurrence external cause code.}
}
\value{
The corresponding place of occurrence external cause code as a number, or an error if the input is invalid.
}
\description{
This function takes one parameter, which can be a number or a string. If it is a number,
the function verifies it against predefined codes. If it is not one of the codes, it returns an error.
If it is a text, the function returns the number corresponding to the code. If it is NULL,
the function returns NULL.
}
\examples{
code_place(1)         # returns 1
code_place("Residential institution") # returns 1
code_place(NULL)       # returns NULL
}
