% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{codify_ans}
\alias{codify_ans}
\title{Recodify yes/no}
\usage{
codify_ans(ans = c("No", "Yes", "Unknowns"))
}
\arguments{
\item{ans}{Codification for an answer as either numeric or text \code{0} (No), \code{1} (Yes) or
\code{9} (Unknown)}
}
\value{
Answer codified as \code{0} (No), \code{1} (Yes) or \code{9} (Unknown)
}
\description{
Gives the codification for a yes/no rule according to DORIS codification
}
\keyword{internal}
