% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{codify_manner_of_death}
\alias{codify_manner_of_death}
\title{Verify and Convert Manner of Death Code}
\usage{
codify_manner_of_death(manner_of_death)
}
\arguments{
\item{manner_of_death}{Either code or word: 0: Disease, 1: Accident, 2: Intentional self-harm,
3: Assault, 4: Legal intervention, 5: War, 6: Could not be determined,
7: Pending investigation, 9: Unknown.}
}
\value{
The corresponding manner of death code as a number, or an error if the input is invalid.
}
\description{
This function takes one parameter, which can be a number or a string. If it is a number,
the function verifies it against predefined codes. If it is not one of the codes, it returns an error.
If it is a text, the function returns the number corresponding to the code. If it is NULL,
the function returns NULL.
}
\keyword{internal}
