% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{codify_sex}
\alias{codify_sex}
\title{Recodify sex}
\usage{
codify_sex(sex = c("Males", "Females", "Unknowns"))
}
\arguments{
\item{sex}{Can be either numeric (\code{1} for \code{Male}, \code{2} for \code{Female} and \code{9} for \code{Unknown}) or can
be a character with those words.}
}
\value{
Sex codified as \code{1}, \code{2} or \code{9}
}
\description{
Gives the codification for sex as Male, Female or Unknown according to DORIS rules
}
\keyword{internal}
