% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doris.R
\name{doris}
\alias{doris}
\title{Interface for the DORIS system of WHO}
\usage{
doris(
  token,
  sex = NULL,
  age = NULL,
  date_of_birth = NULL,
  date_of_death = NULL,
  cause_of_death_code_a = NULL,
  cause_of_death_code_b = NULL,
  cause_of_death_code_c = NULL,
  cause_of_death_code_d = NULL,
  cause_of_death_code_e = NULL,
  cause_of_death_uri_a = NULL,
  cause_of_death_uri_b = NULL,
  cause_of_death_uri_c = NULL,
  cause_of_death_uri_d = NULL,
  cause_of_death_uri_e = NULL,
  cause_of_death_code_part2 = NULL,
  cause_of_death_uri_part2 = NULL,
  interval_a = NULL,
  interval_b = NULL,
  interval_c = NULL,
  interval_d = NULL,
  interval_e = NULL,
  surgery_was_performed = NULL,
  surgery_date = NULL,
  surgery_reason = NULL,
  autopsy_was_requested = NULL,
  autopsy_findings = NULL,
  manner_of_death = NULL,
  manner_of_death_date_of_external_cause_or_poisoning = NULL,
  manner_of_death_description_external_cause = NULL,
  manner_of_death_place_of_occurrence_external_cause = NULL,
  fetal_or_infant_death_multiple_pregnancy = NULL,
  fetal_or_infant_death_stillborn = NULL,
  fetal_or_infant_death_within_24h = NULL,
  fetal_or_infant_death_birth_weight = NULL,
  fetal_or_infant_death_pregnancy_weeks = NULL,
  fetal_or_infant_death_age_mother = NULL,
  fetal_or_infant_death_perinatal_description = NULL,
  maternal_death_was_pregnant = NULL,
  maternal_death_time_from_pregnancy = NULL,
  maternal_death_pregnancy_contribute = NULL,
  release = "2024-01",
  language = "en",
  api_version = "v2"
)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{sex}{Can be either numeric (\code{1} for \code{Male}, \code{2} for \code{Female} and \code{9} for \code{Unknown}) or can
be a character with those words.}

\item{age}{Estimated age of the individual in ISO-8601 format. Use \code{\link[=iso_8601]{iso_8601()}} to create the
age in the appropriate format.}

\item{date_of_birth}{Date of birth of the individual (date or datetime object)}

\item{date_of_death}{Date of death of the individual (date or datetime object)}

\item{cause_of_death_code_a}{String. Cause field A classification codes, comma separated.}

\item{cause_of_death_code_b}{String. Cause field B classification codes, comma separated.}

\item{cause_of_death_code_c}{String. Cause field C classification codes, comma separated.}

\item{cause_of_death_code_d}{String. Cause field D classification codes, comma separated.}

\item{cause_of_death_code_e}{String. Cause field E classification codes, comma separated.}

\item{cause_of_death_uri_a}{String. Cause field A classification URIs, comma separated.}

\item{cause_of_death_uri_b}{String. Cause field B classification URIs, comma separated.}

\item{cause_of_death_uri_c}{String. Cause field C classification URIs, comma separated.}

\item{cause_of_death_uri_d}{String. Cause field D classification URIs, comma separated.}

\item{cause_of_death_uri_e}{String. Cause field E classification URIs, comma separated.}

\item{cause_of_death_code_part2}{String. Cause field Part2 classification codes, comma separated.}

\item{cause_of_death_uri_part2}{String. Cause field Part2 classification URIs, comma separated.}

\item{interval_a}{String. Time interval from onset to death for Field A, in ISO 8601 format.}

\item{interval_b}{String. Time interval from onset to death for Field B, in ISO 8601 format.}

\item{interval_c}{String. Time interval from onset to death for Field C, in ISO 8601 format.}

\item{interval_d}{String. Time interval from onset to death for Field D, in ISO 8601 format.}

\item{interval_e}{String. Time interval from onset to death for Field E, in ISO 8601 format.}

\item{surgery_was_performed}{Whether surgery was performed on the individual
(\code{0} or \code{No}, \code{1} or \code{Yes}, and \code{9} or \code{Unknown})}

\item{surgery_date}{Date of surgery on the individual if \code{surgery_was_performed}}

\item{surgery_reason}{String. Reason for surgery.}

\item{autopsy_was_requested}{Whether an autposy was performed on the individual
(\code{0} or \code{No}, \code{1} or \code{Yes}, and \code{9} or \code{Unknown})}

\item{autopsy_findings}{(\code{0} or \code{No}, \code{1} or \code{Yes}, and \code{9} or \code{Unknown})}

\item{manner_of_death}{Integer. 0: Disease, 1: Accident, 2: Intentional self-harm, 3: Assault,
4: Legal intervention, 5: War, 6: Could not be determined, 7: Pending investigation, 9: Unknown.}

\item{manner_of_death_date_of_external_cause_or_poisoning}{Date of external cause or poisoning of the individual
if \code{manner_of_death} is \code{1}, \code{2}, \code{3}, \code{4}, \code{5}, \code{6},}

\item{manner_of_death_description_external_cause}{String. Description of external cause.}

\item{manner_of_death_place_of_occurrence_external_cause}{Integer. 0: At home, 1: Residential institution, 2: School, other institution, public administration area, 3: Sports and athletics area, 4: Street and highway, 5: Trade and service area, 6: Industrial and construction area, 7: Farm, 8: Other place, 9: Unknown.}

\item{fetal_or_infant_death_multiple_pregnancy}{(\code{0} or \code{No}, \code{1} or \code{Yes}, and \code{9} or \code{Unknown})}

\item{fetal_or_infant_death_stillborn}{(\code{0} or \code{No}, \code{1} or \code{Yes}, and \code{9} or \code{Unknown})}

\item{fetal_or_infant_death_within_24h}{Hours survived if death occurred within 24 hours.}

\item{fetal_or_infant_death_birth_weight}{Birth weight in grams.}

\item{fetal_or_infant_death_pregnancy_weeks}{Integer. Number of completed weeks of pregnancy.}

\item{fetal_or_infant_death_age_mother}{Integer. Age of mother in years.}

\item{fetal_or_infant_death_perinatal_description}{String. Condition of mother affecting the fetus and newborn.}

\item{maternal_death_was_pregnant}{For women, was the deceased pregnant?
(\code{0} or \code{No}, \code{1} or \code{Yes}, and \code{9} or \code{Unknown})}

\item{maternal_death_time_from_pregnancy}{Integer. 0: At time of death, 1: Within 42 days before death, 2: Between 43 days and 1 year before death, 3: One year or more before death, 9: Unknown.}

\item{maternal_death_pregnancy_contribute}{Did pregnancy contribute to death?
(\code{0} or \code{No}, \code{1} or \code{Yes}, and \code{9} or \code{Unknown})}

\item{release}{Release version as listed in
\url{https://icd.who.int/docs/icd-api/SupportedClassifications/#supported-classifications-and-versions}
the default is \code{2024-01}.}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{api_version}{Version of the WHO API you are using. Don't change it unless you know what
you are doing.}
}
\value{
List with the cause of death report
}
\description{
Provides an interface for the DORIS \url{https://icd.who.int/doris}
system by the World Health Organization. Obtains the underlying cause of death using the
rules of the Doris system. The fields provided are fields of a standard death certificate.
}
\details{
All of the parameters are optional. In case of a missing parameter don't use it.
}
\note{
IMPORTANT!. This endpoint is pre-release. The results may have issues and the API may change in
the next releases. Underlying cause of death detection using the rules of the Doris system
(https://icd.who.int/doris) The fields provided are fields of a standard death certificate.
You don't need to provide causeOfDeathUri fields if you provide causeOfDeathCode fields or
vice versa
}
\examples{
#Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {

  #Get the Token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  #Check a 25 year old male with immediate cause 1C8Z
  #due to 1C68.3
  doris(token, sex = 1, age = iso_8601(years = 25), cause_of_death_code_a = "1C8Z",
        cause_of_death_code_b = "1C62.3Z")

}
}
