% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{.icd10_search_recursive}
\alias{.icd10_search_recursive}
\title{Search the ICD-10 for information recursively}
\usage{
.icd10_search_recursive(
  token,
  search_vals,
  release = 2019,
  language = "en",
  auto_update = TRUE,
  k = 0
)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{search_vals}{Either \code{codes}, \code{blocks} or \code{chapter} values to search.}

\item{release}{(\code{character}) ICD-10 release (\code{default = 2019}). Use
\code{\link[=icd10_releases]{icd10_releases()}} to get all available releases.}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}
}
\description{
Search the ICD-10 using the WHO API for information on blocks, chapters or codes.
}
\details{
Searches for either \code{codes}, \code{blocks} or \code{chapters} in a specified \code{release}.
}
\examples{

#Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  #Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search multiple titles
  dbf <- .icd10_search_recursive(token, c("D50.1","I14","XII","D50","D60","D673845"))
  dbf <- .icd10_search_recursive(token, c("XI","XII", NA_character_))
  dbf <- .icd10_search_recursive(token, c("D50-D53", "D50-D53", NA_character_, "D60-D64"))

}
}
\keyword{internal}
