% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10_utils.R
\name{.icd10_search_tidy}
\alias{.icd10_search_tidy}
\title{Search info in a tidy way}
\usage{
.icd10_search_tidy(
  .data,
  colname,
  searchfun,
  token,
  release,
  language,
  dry_run,
  codes_only,
  ...
)
}
\arguments{
\item{.data}{A \code{data.frame} or a \code{data.frame}extension such as a \code{tibble}.}

\item{colname}{(\code{character}) Name of the \code{data.frame} column with the ICD-10 code/block or chapter
to search}

\item{searchfun}{Search function to be utilized}

\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{release}{(\code{character}) ICD-10 release (\code{default = 2019}). Use
\code{\link[=icd10_releases]{icd10_releases()}} to get all available releases.}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{dry_run}{(\code{logical}) If \code{TRUE} will create the request and generate an empty token. This
function is mostly for development.}

\item{codes_only}{Return only the codes without the titles. This option
is faster as it realizes less requests.}

\item{...}{Additional parameters to pass to \code{searchfun}}
}
\value{
A data frame containing all the parent nodes for the block/code/chapter
as well as the original dataframe
}
\description{
Searches for chapter, block or code in a vectorized fashion and returns
into a tidy object (as a \code{mutate} would)
}
\keyword{internal}
