% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10_utils.R
\name{.icd10_site_title}
\alias{.icd10_site_title}
\title{ICD-10 site title}
\usage{
.icd10_site_title(
  token,
  site,
  release = 2019,
  language = "en",
  dry_run = FALSE,
  auto_update = TRUE
)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{site}{Specific link according to the WHO Swagger API
\verb{https://id.who.int/swagger/index.html}}

\item{release}{(\code{character}) ICD-10 release (\code{default = 2019}). Use
\code{\link[=icd10_releases]{icd10_releases()}} to get all available releases.}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{dry_run}{(\code{logical}) If \code{TRUE} will create the request and generate an empty token. This
function is mostly for development.}

\item{auto_update}{Attempts to update token automatically}
}
\value{
A character with the title of the site
}
\description{
Obtain the title of an ICD-10 site
as requested to \verb{https://id.who.int/icd/release/10/\\\{release\\\}/\\\{site\\\}}
}
