% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{icd10_code_search_release}
\alias{icd10_code_search_release}
\title{Search for ICD-10 code across releases}
\usage{
icd10_code_search_release(
  token,
  code,
  language = "en",
  validate_code = TRUE,
  auto_update = TRUE
)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{code}{(\code{character}) ICD-10 code}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{validate_code}{( \code{logical}) Apply additional functions to validate that
the ICD-10 code is written as the API needs it.}

\item{auto_update}{Attempts to update token automatically}
}
\value{
A vector of characters with the ICD-10 releases where you can find
the ICD-10 code
}
\description{
Searches for a specific ICD-10 code across different releases and returns
the editions of those releases in which the code appears.
}
\examples{
# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  #Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Get title and url for release
  icd10_code_search_release(token, "I70")

}

}
