% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{icd10_codes}
\alias{icd10_codes}
\title{Codes of ICD-10 block}
\usage{
icd10_codes(token, block, release = 2019, language = "en", codes_only = FALSE)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{block}{Character with range of values that conform an ICD-10 code (such as \code{"I70-I79"}).
Note that not all possible ranges are blocks. See \code{\link[=icd10_blocks]{icd10_blocks()}} to list all blocks for a
chapter.}

\item{release}{(\code{character}) ICD-10 release (\code{default = 2019}). Use
\code{\link[=icd10_releases]{icd10_releases()}} to get all available releases.}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{codes_only}{Return only the codes without the titles. This option
is faster as it realizes less requests.}
}
\value{
Data frame with all the codes inside a block and their
description
}
\description{
Lists the available codes inside a block for an ICD-10
release.
}
\examples{
#Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  #Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Get title and url for release
  icd10_codes(token, "I70-I79")
}

}
