% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{icd10_name_children}
\alias{icd10_name_children}
\title{ICD-10 chapter, section or code child node names}
\usage{
icd10_name_children(
  token,
  site,
  release = 2019,
  language = "en",
  codes_only = FALSE
)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{site}{Specific link according to the WHO Swagger API
\verb{https://id.who.int/swagger/index.html}}

\item{release}{(\code{character}) ICD-10 release (\code{default = 2019}). Use
\code{\link[=icd10_releases]{icd10_releases()}} to get all available releases.}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{codes_only}{Return only the codes without the titles. This option
is faster as it realizes less requests.}
}
\value{
A \code{data.frame} with the names of all the children (column \code{title})
available at \code{site} as well as their codes (column \code{code}).
}
\description{
Generates a request to
\verb{https://id.who.int/icd/release/10/\\\{release\\\}/\\\{site\\\}},
lists the child nodes and names them.
}
\keyword{internal}
