% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{icd10_title}
\alias{icd10_title}
\title{ICD-10 titles}
\usage{
icd10_title(
  token,
  search_val,
  release = 2019,
  language = "en",
  auto_update = TRUE,
  as_data_frame = TRUE
)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{search_val}{(\code{character}) ICD-10 chapter, block or code for searching.}

\item{release}{(\code{character}) ICD-10 release (\code{default = 2019}). Use
\code{\link[=icd10_releases]{icd10_releases()}} to get all available releases.}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{auto_update}{Attempts to update token automatically}

\item{as_data_frame}{Whether to return the result as a \code{data.frame} or as a vector.}
}
\value{
A character with the title of the chapter, block or code
}
\description{
Obtain the title of an ICD-10 chapter, block or code
}
\examples{
#Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  #Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Get title and url for release
  icd10_title(token, "XII")

  #Do multiple at once
  icd10_title(token, c("XII","IX","XII","II","D50","D50.1"))
}

}
