% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10_utils.R
\name{icd10_validate_code}
\alias{icd10_validate_code}
\title{ICD-10 code validation}
\usage{
icd10_validate_code(code)
}
\arguments{
\item{code}{(\code{character}) ICD-10 code}
}
\value{
A character that is capitalized and has a dot separating the
first 3 digits with the next digits so that it might resemble an ICD-10
code
}
\description{
Takes an ICD-10 code and transforms it using capital letters. It also
checks that it has a point for the last digits (so e141 is transformed to
E14.1)
}
\keyword{internal}
