% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{icd10info}
\alias{icd10info}
\alias{icd10_block_info}
\alias{icd10_code_info}
\alias{icd10_chapter_info}
\title{ICD-10 complete information on code, block or chapter}
\usage{
icd10_block_info(
  token,
  block,
  release = 2019,
  language = "en",
  dry_run = FALSE,
  codes_only = FALSE
)

icd10_code_info(
  token,
  code,
  release = 2019,
  language = "en",
  dry_run = FALSE,
  codes_only = FALSE,
  validate_code = TRUE
)

icd10_chapter_info(
  token,
  chapter,
  release = 2019,
  language = "en",
  dry_run = FALSE,
  codes_only = FALSE
)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{block}{ICD-10 section codes (hyphenated section codes)}

\item{release}{(\code{character}) ICD-10 release (\code{default = 2019}). Use
\code{\link[=icd10_releases]{icd10_releases()}} to get all available releases.}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{dry_run}{(\code{logical}) If \code{TRUE} will create the request and generate an empty token. This
function is mostly for development.}

\item{codes_only}{Return only the codes without the titles. This option
is faster as it realizes less requests.}

\item{code}{(\code{character}) ICD-10 code}

\item{validate_code}{( \code{logical}) Apply additional functions to validate that
the ICD-10 code is written as the API needs it.}

\item{chapter}{(\code{character}) ICD-10 chapter as a roman numeral}
}
\value{
A character with the title of the chapter, block (and code if
code is specified).
}
\description{
Obtains the title of an ICD-10 block or code, as well
as the chapter (and block) to which it belongs (in a certain release).
}
\examples{
# Change `dry_run = FALSE` to run the example
token <- get_token("123", "123", dry_run = TRUE)
icd10_block_info(token, "D60-D64", dry_run = TRUE)

# Change `dry_run = FALSE` to run the example
token <- get_token("123", "123", dry_run = TRUE)
icd10_code_info(token, "D60", dry_run = TRUE)
icd10_code_info(token, "D60.1", dry_run = TRUE)

# Change `dry_run = FALSE` to run the example
token <- get_token("123", "123", dry_run = TRUE)
icd10_chapter_info(token, "II", dry_run = TRUE)

}
