% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd11.R
\name{icd11_lookup}
\alias{icd11_lookup}
\title{Looks up a foundation entity within a linearization}
\usage{
icd11_lookup(
  token,
  id,
  release = "2024-01",
  linearization = "mms",
  language = "en",
  auto_update = TRUE
)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{id}{(stirng) String either with a complete URI of the last digits of the URI}

\item{release}{Release version as listed in
\url{https://icd.who.int/docs/icd-api/SupportedClassifications/#supported-classifications-and-versions}
the default is \code{2024-01}.}

\item{linearization}{(\code{character}) ICD-11 linearization (\code{default = "mms"}). See all linearizations
in the WHO's API website \url{https://icd.who.int/docs/icd-api/APIDoc-Version2/#the-icd11-linearizations}}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{auto_update}{Attempts to update token automatically}
}
\description{
Looks up a foundation entity within a linearization and returns where that entity is coded
in this linearization.
}
\note{
This function is mostly for development of extensions. End-users should refer to
\code{icd_search()}.
}
\examples{
# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-11 URI across multiple releases
  icd11_lookup(token, id = "http://id.who.int/icd/entity/1442995018")
}

}
