% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd11.R
\name{id}
\alias{id}
\alias{icd11_id}
\alias{icd11_id2}
\alias{icd11_id3}
\title{Identify an URI in ICD-11}
\usage{
icd11_id(
  token,
  id,
  linearization = "mms",
  release = "2024-01",
  include = NULL,
  language = "en",
  auto_update = TRUE
)

icd11_id2(
  token,
  id,
  release = "2024-01",
  include = NULL,
  language = "en",
  auto_update = TRUE
)

icd11_id3(
  token,
  id,
  linearization = "mms",
  language = "en",
  auto_update = TRUE
)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{id}{(stirng) String either with a complete URI of the last digits of the URI}

\item{linearization}{(\code{character}) ICD-11 linearization (\code{default = "mms"}). See all linearizations
in the WHO's API website \url{https://icd.who.int/docs/icd-api/APIDoc-Version2/#the-icd11-linearizations}}

\item{release}{Release version as listed in
\url{https://icd.who.int/docs/icd-api/SupportedClassifications/#supported-classifications-and-versions}
the default is \code{2024-01}.}

\item{include}{Either NULL of provide additional information such as \code{ancestor},
\code{decendant} or \code{diagnosticCriteria}. For example, if you set \code{include="ancestor"} the returned
response will include all ancestor entity URIs for the requested entity. Use a vector to include
multiple properties. For example \code{include = c("ancestor", "diagnosticCriteria")}}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{auto_update}{Attempts to update token automatically}
}
\value{
Character vector with the ICD-11 URI and its characteristics
}
\description{
Provides all of the characteristics from an URI
}
\details{
\code{icd_id()} Searches for a specific URI across a specific \code{linearization} an \code{release}.
\code{icd_id2()}  searches for a URI across a specific release regardless of linearization. Finally,
\code{icd_id3} searches for the URI across all ICD-11 releases.
}
\examples{
# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-11 URI
  icd11_id(token, id = "http://id.who.int/icd/entity/1442995018")

  # Search for ICD-11 URI with parents and diagnostic criteria
  icd11_id(token,
    id = "http://id.who.int/icd/entity/1442995018",
    include = c("ancestor", "diagnosticCriteria")
  )
}
# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-11 URI across all linearizations in a specific release
  icd11_id2(token, id = "http://id.who.int/icd/entity/1442995018")

  # Search for ICD-11 URI with parents and diagnostic criteria across all linearizations
  #  in a specific release
  icd11_id2(token,
    id = "http://id.who.int/icd/entity/1442995018",
    include = c("ancestor", "diagnosticCriteria")
  )
}
# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-11 URI across multiple releases
  icd11_id3(token, id = "http://id.who.int/icd/entity/1442995018")

}

}
