% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iso_8601}
\alias{iso_8601}
\title{Generate a duration period for the DORIS system in ISO 8601 format}
\usage{
iso_8601(
  years = 0,
  months = 0,
  weeks = 0,
  days = 0,
  hours = 0,
  minutes = 0,
  seconds = 0
)
}
\arguments{
\item{years}{Number of years in the duration (default = 0).}

\item{months}{Number of months in the duration (default = 0).}

\item{weeks}{Number of weeks in the duration (default = 0).}

\item{days}{Number of days in the duration (default = 0).}

\item{hours}{Number of hours in the duration (default = 0).}

\item{minutes}{Number of minutes in the duration (default = 0).}

\item{seconds}{Number of seconds in the duration (default = 0).}
}
\value{
A time duration formatted according to ISO-8601
}
\description{
Creates a duration period for the DORIS system in ISO-8601 format
from inputting years, months, weeks, days, hours, minutes, and seconds
}
\examples{

#A duration with 24 years
iso_8601(years = 24)

#A duration with 6 months, 2 weeks and 5 days
iso_8601(months = 6, weeks = 2, days = 5)

}
