% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_token.R
\name{make_request}
\alias{make_request}
\title{Make a request for the WHO API v2}
\usage{
make_request(url, token, language = "en", auto_update = TRUE, req_body = NULL)
}
\arguments{
\item{url}{(\code{character}) Address for the request}

\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{auto_update}{Attempts to update token automatically}
}
\value{
An \code{httr2} request for the API
}
\description{
Builds a request to the \code{url} using the \code{token} obtained
from the \code{\link[=get_token]{get_token()}} function.
}
\examples{
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  #Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  res <- make_request(
      url = 'https://id.who.int/icd/release/10/E14.1',
      token = token,
  )
  #Use that request with the API:
  res |> req_perform()
 }

}
\keyword{internal}
