% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_df}
\alias{name_df}
\title{Function to add names to chapters / blocks to df}
\usage{
name_df(df)
}
\arguments{
\item{df}{A data frame with columns \code{level_0}, \code{title_0}, \code{level_1}, \code{title_1}, etc.}
}
\value{
A named \code{data.frame}
}
\description{
Adds the appropriate names to chapters, blocks and codes
}
\keyword{internal}
