% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_columns}
\alias{order_columns}
\title{Function to order the column names of a searched data frame}
\usage{
order_columns(df)
}
\arguments{
\item{df}{A data frame with columns \code{level_0}, \code{title_0}, \code{level_1}, \code{title_1}, etc.}
}
\value{
The dataframe with the columns ordered according to level number.
}
\description{
Function to order the column names of a searched data frame
}
\keyword{internal}
