% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_token.R
\name{request_WHO}
\alias{request_WHO}
\title{Create and perform a request to the WHO API v2.}
\usage{
request_WHO(
  url,
  token = NULL,
  language = "en",
  auto_update = TRUE,
  warning_message_404 = "url for request not found",
  post_process_function = function(x) {
     x
 }
)
}
\arguments{
\item{url}{(\code{character}) Address for the request}

\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{auto_update}{Attempts to update token automatically}

\item{post_process_function}{Function to apply to the body of the response after
obtaining a response.}
}
\value{
The response's body from the request
}
\description{
This is the main function for development. Creates a request for the API using \code{make_request}
and then performs the request using \code{httr2::req_perform}. Finally it obtains the body of the
response and returns to user.
}
\examples{
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {

  #Get the URL
  url <- "https://id.who.int/icd/release/10"

  #Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Get title and url for release
  request_WHO(url, token = token)
}
}
\keyword{internal}
