% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd11.R
\name{residual}
\alias{residual}
\alias{icd11_residual}
\alias{icd11_residual2}
\title{Residual}
\usage{
icd11_residual(
  token,
  id,
  residual = "",
  linearization = "mms",
  release = "2024-01",
  include = NULL,
  language = "en",
  auto_update = TRUE
)

icd11_residual2(
  token,
  id,
  residual = "",
  linearization = "mms",
  include = NULL,
  language = "en",
  auto_update = TRUE
)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{id}{(stirng) String either with a complete URI of the last digits of the URI}

\item{residual}{For residual categories: - this could be 'other' for other specified residual
category - or 'unspecified' for the unspecified residual category}

\item{linearization}{(\code{character}) ICD-11 linearization (\code{default = "mms"}). See all linearizations
in the WHO's API website \url{https://icd.who.int/docs/icd-api/APIDoc-Version2/#the-icd11-linearizations}}

\item{release}{Release version as listed in
\url{https://icd.who.int/docs/icd-api/SupportedClassifications/#supported-classifications-and-versions}
the default is \code{2024-01}.}

\item{include}{Either NULL of provide additional information such as \code{ancestor},
\code{decendant} or \code{diagnosticCriteria}. For example, if you set \code{include="ancestor"} the returned
response will include all ancestor entity URIs for the requested entity. Use a vector to include
multiple properties. For example \code{include = c("ancestor", "diagnosticCriteria")}}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{auto_update}{Attempts to update token automatically}
}
\description{
Returns the residual from a URI
}
\examples{
# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-11 URI across multiple releases
  icd11_residual(token, id = "1442995018")

}
# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-11 URI across multiple releases
  icd11_residual2(token, id = "1442995018")

}
}
