% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve}
\alias{retrieve}
\title{Return list entries from search}
\usage{
retrieve(list_val, val)
}
\arguments{
\item{lname}{Name of the list entry to recover as a vector}
}
\value{
A
}
\description{
Function that retrieves a list of function
}
\examples{
val <- "title"
search_results <- icd11_search(token, "HIV stage 4")
search_results |> retrieve("title")
}
