% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{right_align_df}
\alias{right_align_df}
\title{Function to right align the data.frame of chapters}
\usage{
right_align_df(df)
}
\arguments{
\item{df}{A data frame with columns \code{level_0}, \code{title_0}, \code{level_1}, \code{title_1}, etc.}
}
\value{
A right-aligned \code{data.frame}
}
\description{
Right aligns the search data frame so that it has the correct levels
}
\keyword{internal}
