% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_request}
\alias{run_request}
\title{Run a request for the WHO API}
\usage{
run_request(res, dry_run)
}
\arguments{
\item{res}{An \code{httr2} request to perform}
}
\value{
Either the result from the request or \code{NULL} if \code{dry_run} is \code{TRUE}
}
\description{
Runs a request \code{res} if \code{dry_run} is not \code{FALSE}. Returns
\code{NULL} otherwise.
}
\keyword{internal}
