% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd11.R
\name{search}
\alias{search}
\alias{icd11_search}
\alias{icd11_search2}
\title{Search}
\usage{
icd11_search(
  token,
  text,
  release = "2024-01",
  linearization = "mms",
  subtrees_filter = NULL,
  subtrees_filter_foundation_descendents = FALSE,
  include_keyword_result = FALSE,
  chapter_filter = NULL,
  use_flexisearch = FALSE,
  flat_results = NULL,
  highlighting_enabled = FALSE,
  medical_coding_mode = TRUE,
  properties_to_be_searched = NULL,
  language = "en",
  auto_update = TRUE
)

icd11_search2(
  token,
  text,
  release = "2024-01",
  linearization = "mms",
  subtrees_filter = NULL,
  chapter_filter = NULL,
  use_flexisearch = FALSE,
  flat_results = NULL,
  highlighting_enabled = FALSE,
  properties_to_be_searched = NULL,
  language = "en",
  auto_update = TRUE
)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{text}{The text component to search in ICD-11.}

\item{release}{Release version as listed in
\url{https://icd.who.int/docs/icd-api/SupportedClassifications/#supported-classifications-and-versions}
the default is \code{2024-01}.}

\item{linearization}{(\code{character}) ICD-11 linearization (\code{default = "mms"}). See all linearizations
in the WHO's API website \url{https://icd.who.int/docs/icd-api/APIDoc-Version2/#the-icd11-linearizations}}

\item{subtrees_filter}{(string) Optional parameter. Vector of URIs. If provided,
the search will be performed on the entities provided and their descendants}

\item{subtrees_filter_foundation_descendents}{When a \code{subtrees_filter} is used the search looks at
the linearization descendants of the entities in the \code{subtrees_filter}. If

\code{subtrees_filter_foundation_descendents} is set to \code{TRUE}, the search looks at the
foundation descendants of the entities in the \code{subtrees_filter}. Default value: \code{FALSE}}

\item{include_keyword_result}{Boolean. Optional parameter. If set to \code{TRUE}, the search result
will also include keyword list. Default value: \code{FALSE}}

\item{chapter_filter}{Vector of chapter codes e.g., c("01","02","21").
When provided, the search will be performed only on these chapters.}

\item{use_flexisearch}{Boolean. Changes the search mode to flexible search. Default value: \code{FALSE}}

\item{flat_results}{Boolean. Optional parameter. If set to true the search result entities are
provided in a nested data structure representing the ICD-11 hierarchy. Default value: true.}

\item{highlighting_enabled}{Boolean. Optional. If set to \code{FALSE} the search result highlighting
is turned off. Default value: \code{FALSE}. Highlighting is in html format.}

\item{medical_coding_mode}{Boolean. This mode is the default and should be used when searching
the classification for medical coding purposes. Default value: \code{TRUE}.}

\item{properties_to_be_searched}{String. The properties to be included in the search.
Used only when medical_coding_mode=\code{FALSE}}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{auto_update}{Attempts to update token automatically}
}
\value{
A list of the coded information as well as any additional information.
}
\description{
Searches the foundation component of the ICD-11.
}
\details{
The \code{icd11_search} function allows you to search across different linearizations such
as the Mortality and Morbidity Statistics \code{linearization = "mms"} (default) and the
International Classifications of Functioning Disability and Health \code{linearization = "icf"}.
}
\note{
Both \code{icd11_search()} and \code{icd11_search2()} provide options for search. They
correspond to two different API entry-points in the WHO's API and provide slighly different
amount of information for the same result.

This provides the default search but might throw different amount of information
as \code{icd11_search2}.
}
\examples{
# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-10 releases
  search_results <- icd11_search(token, text = "HIV stage 4")
}

# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-10 releases
  search_results <- icd11_search2(token, text = "Cerebrovascular accident")
}

}
