% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{tidysearch}
\alias{tidysearch}
\title{Search for information on multiple codes, blocks or chapter in a
\code{data.frame} or \code{tibble}.}
\arguments{
\item{.data}{A \code{data.frame} or a \code{data.frame}extension such as a \code{tibble}.}

\item{var}{Name of the \code{data.frame} column with the ICD-10 code that
requires the info search}
}
\description{
Search for information on multiple codes, blocks or chapter in a
\code{data.frame} or \code{tibble}.
}
