% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{vectorsearch}
\alias{vectorsearch}
\alias{icd10_code_info_vectorized}
\alias{icd10_block_info_vectorized}
\alias{icd10_chapter_info_vectorized}
\title{Search for information on multiple codes, blocks or chapter in a
\code{vector}.}
\usage{
icd10_code_info_vectorized(
  token,
  codes,
  release = 2019,
  language = "en",
  dry_run = FALSE,
  codes_only = FALSE,
  validate_code = TRUE
)

icd10_block_info_vectorized(
  token,
  blocks,
  release = 2019,
  language = "en",
  dry_run = FALSE,
  codes_only = FALSE
)

icd10_chapter_info_vectorized(
  token,
  chapters,
  release = 2019,
  language = "en",
  dry_run = FALSE,
  codes_only = FALSE
)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{codes}{A \code{vector} containing multiple codes.}

\item{release}{(\code{character}) ICD-10 release (\code{default = 2019}). Use
\code{\link[=icd10_releases]{icd10_releases()}} to get all available releases.}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{dry_run}{(\code{logical}) If \code{TRUE} will create the request and generate an empty token. This
function is mostly for development.}

\item{codes_only}{Return only the codes without the titles. This option
is faster as it realizes less requests.}

\item{validate_code}{( \code{logical}) Apply additional functions to validate that
the ICD-10 code is written as the API needs it.}

\item{blocks}{A \code{vector} containing multiple blocks.}

\item{chapters}{A \code{vector} containing multiple chapters}
}
\value{
A \code{data.frame} with information for each code in \code{codevec}
}
\description{
Search for information on multiple codes, blocks or chapter in a
\code{vector}.
}
\note{
All entries in the vector must be of the same level. That is, you
cannot combine codes with blocks or codes with chapters. If you wish
to do so generate two different vectors one for the codes and one for the
chapters
}
\examples{
# Change `dry_run = FALSE` to run the example
token <- get_token("123", "123", dry_run = TRUE)
icd10_code_info_vectorized(token,
  codes = c("E14.1", "C80.0", "F14"),
  dry_run = TRUE
)
# Change `dry_run = FALSE` to run the example
token <- get_token("123", "123", dry_run = TRUE)
icd10_block_info_vectorized(token,
  blocks = c("E10-E14", "F10-F19", "C76-C80"),
  dry_run = TRUE
)
# Change `dry_run = FALSE` to run the example
token <- get_token("123", "123", dry_run = TRUE)
icd10_chapter_info_vectorized(token,
  chapters = c("XII", "II", "V"),
  dry_run = TRUE
)
}
