% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_token_expiration_time}
\alias{check_token_expiration_time}
\title{Check token expiration time}
\usage{
check_token_expiration_time(token)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}
}
\value{
NULL It checks whether the token has expired or is soon to expire
and alerts the user.
}
\description{
Checks whether the token will expire soon or has already expired and
throws a warning to the user.
}
\keyword{internal}
