% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_token.R
\name{get_token}
\alias{get_token}
\title{Request an access token from ICD API}
\usage{
get_token(client_id, client_secret, dry_run = FALSE)
}
\arguments{
\item{client_id}{(\code{character}) The client ID as provided by the WHO website (see \strong{Generating credentials}).}

\item{client_secret}{(\code{character}) The client secret key as provided by the WHO API website (see \strong{Generating credentials}).}

\item{dry_run}{(\code{logical}) If \code{TRUE} will create the request and generate an empty token. This
function is mostly for development.}
}
\value{
Access token for the WHO API. \strong{Don't share this token!!}
}
\description{
Obtains an access token for the World Health Organization's (WHO) International
Classification of Diseases (ICD) API available at
\url{https://icd.who.int/icdapi}.
}
\section{Generating credentials}{


Go to the World Health Organization API website
\url{https://icd.who.int/icdapi}
and \strong{Register} in the \strong{API Access} section. Once you have registered and
accepted the terms and conditions click on \verb{View API access key(s)}.

You will get a \code{ClientId} and a \code{ClientSecret}. Those strings are the ones
required to use the WHO API.
}

\examples{
#Substitute CLIENT_ID and CLIENT_SECRET for your ID and secret
CLIENT_ID     <- Sys.getenv("CLIENT_ID")
CLIENT_SECRET <- Sys.getenv("CLIENT_SECRET")
token <- get_token(CLIENT_ID, CLIENT_SECRET, dry_run = TRUE)
}
