% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_token.R
\name{request_WHO}
\alias{request_WHO}
\title{Create and perform a request to the WHO API v2.}
\usage{
request_WHO(
  url,
  token = NULL,
  language = "en",
  auto_update = TRUE,
  dry_run = FALSE,
  warning_message_404 = "url for request not found",
  post_process_function = function(x) {
     x
 }
)
}
\arguments{
\item{url}{(\code{character}) Address for the request}

\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{auto_update}{Attempts to update token automatically}

\item{post_process_function}{Function to apply to the body of the response after
obtaining a response.}
}
\value{
Either the response from the request or \code{NULL} if \code{dry_run} is \code{TRUE}
}
\description{
This is the main function for development. Creates a request for the API using \code{make_request}
and then performs the request using \code{run_request}. Finally it obtains the body of the
response and returns to user.
}
\examples{

#Get the URL
url <- "https://id.who.int/icd/release/10"

#Obtain your token using your id and secret
token <- get_token("123", "123", dry_run = TRUE)

#Run the request
request_WHO("https://id.who.int/icd/release/10", token, dry_run = TRUE)
}
\keyword{internal}
