% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve}
\alias{retrieve}
\title{Return list entries from search}
\usage{
retrieve(list_val, val)
}
\arguments{
\item{list_val}{List to extract the entry \code{val} from}

\item{val}{Name of the list entry to recover as a vector}
}
\value{
The entry \code{val} from \code{list_val} as vector
}
\description{
Function that retrieves a list of function
}
\examples{
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)
  val <- "title"
  search_results <- icd11_search(token, "HIV stage 4")
  search_results |> retrieve("title")
}
}
