% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_token.R
\name{run_request}
\alias{run_request}
\title{Perform a request for the WHO API}
\usage{
run_request(res, dry_run = FALSE)
}
\arguments{
\item{res}{An \code{httr2} request to perform}
}
\value{
Either the result from the request or \code{NULL} if \code{dry_run} is \code{TRUE}
}
\description{
Performs a request \code{res} if \code{dry_run} is not \code{FALSE}. Returns
\code{NULL} otherwise.
}
\examples{
#Create the request and perform it
res <- make_request(
    url = 'https://id.who.int/icd/release/10/E14.1',
    token = token,
)
#Use that request with the API:
res |> run_request()
}
\keyword{internal}
