% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{titles}
\alias{titles}
\alias{icd10_chapter_title}
\alias{icd10_block_title}
\alias{icd10_code_title}
\title{ICD-10 titles}
\usage{
icd10_chapter_title(
  token,
  chapter,
  release = 2019,
  language = "en",
  dry_run = FALSE
)

icd10_block_title(
  token,
  block,
  release = 2019,
  language = "en",
  dry_run = FALSE
)

icd10_code_title(
  token,
  code,
  release = 2019,
  language = "en",
  dry_run = FALSE,
  validate_code = TRUE
)
}
\arguments{
\item{token}{(\code{list}) Access token to the API obtained from \code{\link[=get_token]{get_token()}}}

\item{chapter}{(\code{character}) ICD-10 chapter as a roman numeral}

\item{release}{(\code{character}) ICD-10 release (\code{default = 2019}). Use
\code{\link[=icd10_releases]{icd10_releases()}} to get all available releases.}

\item{language}{(\code{character}) Language for the request. Available languages are
listed in \url{https://icd.who.int/docs/icd-api/SupportedClassifications/}
the default is English (\code{language = "en"}).}

\item{dry_run}{(\code{logical}) If \code{TRUE} will create the request and generate an empty token. This
function is mostly for development.}

\item{block}{ICD-10 section codes (hyphenated section codes)}

\item{code}{(\code{character}) ICD-10 code}

\item{validate_code}{( \code{logical}) Apply additional functions to validate that
the ICD-10 code is written as the API needs it.}
}
\value{
A character with the title of the chapter, block or code
}
\description{
Obtain the title of an ICD-10 chapter, block or code
}
\examples{
#Change `dry_run = FALSE` to run the example
token <- get_token("123","123", dry_run = TRUE)
icd10_chapter_title(token, "XII", dry_run = TRUE)

#Change `dry_run = FALSE` to run the example
token <- get_token("123","123", dry_run = TRUE)
icd10_block_title(token, "L50-L54", dry_run = TRUE)

#Change `dry_run = FALSE` to run the example
token <- get_token("123","123", dry_run = TRUE)
icd10_code_title(token, "E14.1", dry_run = TRUE)

}
